Minor fixes and refactoring
