+++ src/main/java/net/maritimeconnectivity/serviceregistry/models/domain/Instance.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import net.maritimeconnectivity.serviceregistry.utils.GeometryJSONConverter;
 import net.maritimeconnectivity.serviceregistry.utils.GeometryJSONDeserializer;
 import net.maritimeconnectivity.serviceregistry.utils.GeometryJSONSerializer;
+import net.maritimeconnectivity.serviceregistry.utils.ServiceStatusBridge;
 import org.apache.lucene.analysis.core.LowerCaseFilterFactory;
 import org.efficiensea2.maritime_cloud.service_registry.v1.servicespecificationschema.ServiceStatus;
 import org.hibernate.annotations.Cache;
